% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{all_questions_likelihood}
\alias{all_questions_likelihood}
\title{Finds the log-likelihood for the scores acheived by a student on all questions}
\usage{
all_questions_likelihood(ability, discs, diffs, cs, scores)
}
\arguments{
\item{ability}{The ability of the student answering the questions}

\item{discs}{Vector of discriminations of the questions}

\item{diffs}{Vector of difficulties of the questions}

\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}
}
\value{
The log-likelihood
}
\description{
Finds the log-likelihood for the scores acheived by a student on all questions
}
