% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{all_students_likelihood}
\alias{all_students_likelihood}
\title{Finds the log-likelihood from all students answering all questions}
\usage{
all_students_likelihood(abils, discs, diffs, cs, scores)
}
\arguments{
\item{abils}{Vector of abilities of the students answering the questions}

\item{discs}{Vector of discriminations of the questions}

\item{diffs}{Vector of difficulties of the questions}

\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}
}
\value{
The log-likelihood
}
\description{
Finds the log-likelihood from all students answering all questions
}
