% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{check_ests}
\alias{check_ests}
\title{Estimates the bias, variance, and mean squared error of a data frame of estimates}
\usage{
check_ests(est_vals, true_vals)
}
\arguments{
\item{est_vals}{A data frame with the scaled estimated values}

\item{true_vals}{A data frame with the true values}
}
\value{
A list with the estimated bias, variance, and MSE
}
\description{
Estimates the bias, variance, and mean squared error of a data frame of estimates
}
