% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{est_bias}
\alias{est_bias}
\title{Estimates the bias of a data frame of estimates}
\usage{
est_bias(true_vals, est_vals)
}
\arguments{
\item{true_vals}{A data frame with the true values}

\item{est_vals}{A data frame with the scaled estimated values}
}
\value{
The estimated bias
}
\description{
Estimates the bias of a data frame of estimates
}
