% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{est_MSE}
\alias{est_MSE}
\title{Estimates the mean square error of a data frame of estimates}
\usage{
est_MSE(true_vals, est_vals)
}
\arguments{
\item{true_vals}{A data frame with the true values}

\item{est_vals}{A data frame with the scaled estimated values}
}
\value{
The estimated MSE
}
\description{
Estimates the mean square error of a data frame of estimates
}
