% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{estimate}
\alias{estimate}
\title{A single function to call one of the 3 optimisation methods}
\usage{
estimate(data_set, estimate_type = 1, n = 1000, tuning = 0.01, group = T)
}
\arguments{
\item{data_set}{A data frame containing some generated data for a group of students}

\item{estimate_type}{1 for random walk, 2 for optim, 3 for brute force}

\item{n}{Number of iterations for the random walk method}

\item{tuning}{The tuning parameter for the random walk method}

\item{group}{The group parameter for the random walk/optim methods}
}
\value{
A list with the scaled parameter estimates
}
\description{
A single function to call one of the 3 optimisation methods
}
