% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{find_all_optimised_1}
\alias{find_all_optimised_1}
\title{Uses random walk method with n iterations to find optimised question params and abilities}
\usage{
find_all_optimised_1(cs, scores, ns, nq, n, tuning = 0.01, group = T)
}
\arguments{
\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}

\item{ns}{The number of students}

\item{nq}{The number of unique questions}

\item{n}{The number of iterations to use}

\item{tuning}{The standard deviation of the adjustments at each step}

\item{group}{Boolean - True if we should adjust all question parameters at the same time}
}
\value{
A list of the optimised estimates for the parameters
}
\description{
Uses random walk method with n iterations to find optimised question params and abilities
}
