% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{find_all_optimised_2}
\alias{find_all_optimised_2}
\title{Uses optim to find optimised question params and abilities}
\usage{
find_all_optimised_2(cs, scores, ns, nq, group = F)
}
\arguments{
\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}

\item{ns}{The number of students}

\item{nq}{The number of unique questions}

\item{group}{Boolean - False if we want to maximise the profile log-likelihood, True if we want to maximise the full log-likelihood}
}
\value{
A list of the optimised estimates for the parameters
}
\description{
Uses optim to find optimised question params and abilities
}
