% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{find_all_optimised_3}
\alias{find_all_optimised_3}
\title{Checks a grid of all all values in range for discrimination and difficulty}
\usage{
find_all_optimised_3(cs, scores, ns, nq)
}
\arguments{
\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}

\item{ns}{The number of students}

\item{nq}{The number of unique questions}
}
\value{
A list of the optimised estimates for the parameters
}
\description{
Checks a grid of all all values in range for discrimination and difficulty
}
