% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{find_no_abil}
\alias{find_no_abil}
\title{Given a question will calculate an estimate of the probability of getting the question correct through random guessing}
\usage{
find_no_abil(q, n)
}
\arguments{
\item{q}{A question}

\item{n}{A integer giving the sample size to use for the estimation}
}
\value{
A probability of getting the question correct through random guessing
}
\description{
Given a question will calculate an estimate of the probability of getting the question correct through random guessing
}
