% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{find_optimised_abilities}
\alias{find_optimised_abilities}
\title{Finds estimates for all abilities given question parameters}
\usage{
find_optimised_abilities(discs, diffs, cs, scores, ns, nq)
}
\arguments{
\item{discs}{Vector of discriminations of the questions}

\item{diffs}{Vector of difficulties of the questions}

\item{cs}{Vector of the chances of randomly getting the questions correct}

\item{scores}{The scores achieved on the questions by the student}

\item{ns}{The number of students}

\item{nq}{The number of unique questions}
}
\value{
A list of the optimised abilities and the total log-likelihood for these estimates
}
\description{
Finds estimates for all abilities given question parameters
}
