% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{find_test_qs_rand}
\alias{find_test_qs_rand}
\title{Given a test will find the probabity of getting each question correct through random guessing}
\usage{
find_test_qs_rand(t)
}
\arguments{
\item{t}{The test to be updated}
}
\value{
The updated test
}
\description{
Given a test will find the probabity of getting each question correct through random guessing
}
