% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_ran_question}
\alias{gen_ran_question}
\title{Will generate a random question}
\usage{
gen_ran_question(
  name,
  partial,
  min_neg,
  max_opt,
  max_score,
  set_disc = NA,
  set_diff = NA
)
}
\arguments{
\item{name}{The name of the question}

\item{partial}{Whether the question can have partial marks}

\item{min_neg}{The most points that can be subtracted for getting the question incorrect}

\item{max_opt}{The maximum number of options for the question}

\item{max_score}{The highest score that can be achieved on the question for answering correctly}

\item{set_disc}{Used to force a specific discrimination value}

\item{set_diff}{Used to force a specific difficulty value}
}
\value{
The generated question
}
\description{
Will generate a random question
}
