% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_res_from_mod}
\alias{gen_res_from_mod}
\title{Given a student and a test will simulate the result the student acheived when answering the question}
\usage{
gen_res_from_mod(s, q)
}
\arguments{
\item{s}{A student that will be answering the question}

\item{q}{A question that will be answered by the student}
}
\value{
The simulated mark achieved by the student
}
\description{
Given a student and a test will simulate the result the student acheived when answering the question
}
