% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_stand_ran_test}
\alias{gen_stand_ran_test}
\title{Will generate a test}
\usage{
gen_stand_ran_test(n, name, can_neg, dupe = -1)
}
\arguments{
\item{n}{The number of questions on the test}

\item{name}{The name of the test}

\item{can_neg}{Boolean - whether you allow for a negative mark to be achieved overall on the test}

\item{dupe}{How many unique questions should the test have (-1 for all unique)}
}
\value{
The generated test
}
\description{
Will generate a test
}
