% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_student_data}
\alias{gen_student_data}
\title{Given a test and a vector of students, will simulate the results the students received}
\usage{
gen_student_data(t, students, ran_type = "u", need_rand = F)
}
\arguments{
\item{t}{The test to be answered by the students}

\item{students}{A vector of students to take the test or the number of students to be generated}

\item{ran_type}{'u' or 'n', the distribution of abilities of students if you are generating students}

\item{need_rand}{Boolean - do we need to find the random probability of getting the question correct for the questions on the test}
}
\value{
A data frame containing the results from the student taking the test
}
\description{
Given a test and a vector of students, will simulate the results the students received
}
