% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gen_test_res_from_mod}
\alias{gen_test_res_from_mod}
\title{Given a test and a student will simulate the results the student received}
\usage{
gen_test_res_from_mod(t, s, sname)
}
\arguments{
\item{t}{The test to be answered by the student}

\item{s}{The student to take the test}

\item{sname}{A name used to identify the student}
}
\value{
A data frame containing the results from the student taking the test
}
\description{
Given a test and a student will simulate the results the student received
}
