% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCI.R
\name{getAvCI}
\alias{getAvCI}
\title{Will get the confidence interval of the average mark achieved}
\usage{
getAvCI(marks)
}
\arguments{
\item{marks}{A vector containing the marks of all the students}
}
\value{
The confidence interval for average mark
}
\description{
Will get the confidence interval of the average mark achieved
}
