% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{normalise_results}
\alias{normalise_results}
\title{Scales the values of estimates so that the first discrimination is correct}
\usage{
normalise_results(estimates, disc0)
}
\arguments{
\item{estimates}{A list with the estimated values}

\item{disc0}{The true value of discrimination for the first question}
}
\value{
A list of the scaled estimates for the parameters
}
\description{
Scales the values of estimates so that the first discrimination is correct
}
