% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{plot_results}
\alias{plot_results}
\title{Create a plot of the true values against the estimated values}
\usage{
plot_results(normed_ests, true_values)
}
\arguments{
\item{normed_ests}{A data frame with the scaled estimated values}

\item{true_values}{A data frame with the true values}
}
\description{
Create a plot of the true values against the estimated values
}
