% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_estimates.R
\name{question_likelihood}
\alias{question_likelihood}
\title{Finds the log-likelihood for a score being achieved on a question}
\usage{
question_likelihood(ability, a, b, c, score)
}
\arguments{
\item{ability}{The ability of the student answering the question}

\item{a}{The discrimination of the question}

\item{b}{The difficutly of the question}

\item{c}{The chance of randomly getting the question correct}

\item{score}{The score achieved on the question by the student}
}
\value{
The log-likelihood
}
\description{
Finds the log-likelihood for a score being achieved on a question
}
